requirements.txt
asgiref
CacheControl
cachetools
certifi
cffi
charset-normalizer
cryptography
defusedxml
dj-database-url
Django
django-crispy-forms
django-mssql-backend
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
gunicorn
httplib2
idna
msgpack
mysql-connector-python
mysqlclient
oauthlib
packaging
Pillow
proto-plus
pyasn1
pyasn1-modules
pycparser
PyJWT
pyodbc
pyparsing
python3-openid
pytz
requests
requests-oauthlib
rsa
six
social-auth-app-django
social-auth-core
sqlparse
typing_extensions
tzdata
uritemplate
urllib3
firebase-admin
